Add files via upload
Species has a one to many relationship with animals be cause lots of animals can have the same species, but one animal can't have multiple species they belong to. Adoptions has a one to one relationship with Animals because animals can only be adopted once and that instance of adoption can only belong to the one animal. Owners has a one to many relationship with animals because and owner can adopt multiple animals but an animal can't be adopted by separate parties.